// web/src/generated/models/matterSavedSearchCreate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */
import type { MatterSavedSearchCreateVisibility } from './matterSavedSearchCreateVisibility';
import type { MatterSearchRequest } from './matterSearchRequest';

export interface MatterSavedSearchCreate {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  description?: string | null;
  visibility?: MatterSavedSearchCreateVisibility;
  search: MatterSearchRequest;
  /** @maxItems 1000 */
  shared_user_ids?: string[];
}